+++ src/beshell/repl/REPLModule.hpp
@@ -7,6 +7,8 @@
 namespace be{
 
     class BeShell ;
+    class REPLChannel ;
+    class Package ;
     
 	typedef void (*NativeModuleExportorFunc)(JSContext * ctx, NativeModule * module) ;
 
@@ -24,8 +26,15 @@ namespace be{
         static JSValue disableCrypto(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) ;
         static JSValue setCryptoKey(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) ;
         static JSValue setPassword(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) ;
+        static JSValue enableUnknownPkgEvent(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) ;
+        static JSValue disableUnknownPkgEvent(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) ;
+
+        static REPLModule * singleton ;
 
         void exports(JSContext *ctx) ;
+
+    private:
+        static void unknownPkgRelay(REPLChannel * ch, std::unique_ptr<Package> pkg) ;
 
     } ;
 }