+++ changelog.md
@@ -4,6 +4,7 @@ New Feature:
 
 * example 可从 CMakeLists.private.txt 文件中加载 components 目录
 * 完善了 JSDoc 文档
+* repl.enableUnknownPkgEvent() / repl.disableUnknownPkgEvent() — 监听 pkg.unknown 事件
 
 # v1.0.10